+++ tests/testthat/test-arg.R
@@ -1,7 +1,7 @@
 context("arg")
 
 test_that("deprecated() returns the missing argument", {
-  fn <- function(foo = deprecated()) is_missing(foo)
-  expect_true(fn())
-  expect_false(fn(1))
+  fn <- function(foo = deprecated()) is_supplied(foo)
+  expect_false(fn())
+  expect_true(fn(1))
 })