+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # lifecycle 0.1.0.9000
 
+* New `is_supplied()` function to test whether the caller has supplied a
+  `deprecated()` function.
+
 
 # lifecycle 0.1.0